refactor(cli): consolidate command catalog metadata
Why:
- CLI registration and pre-dispatch metadata were maintained in separate catalogs, which allowed command paths to drift between framework registration and JSON error handling.

What Changed:
- Moved command path metadata, root command checks, result command resolution, and subcommand validation data into UcliCommandCatalog.
- Updated pre-dispatch and parse-error policies to use UcliCommandCatalog as the single source.
- Strengthened the help contract test to compare registered help command paths bidirectionally with the catalog.

Impact:
- Public command names and JSON envelopes remain unchanged, while catalog drift is caught by tests.
