+++ tests/Ucli.Tests/Hosting/Cli/CliCommandRegistrationContractTests.cs
@@ -1,6 +1,6 @@
 using MackySoft.Tests;
-using MackySoft.Ucli.Hosting.Cli.Common.Catalog;
 using MackySoft.Ucli.Hosting.Cli.Common.Contracts;
+using MackySoft.Ucli.Hosting.Cli.Common.Startup;
 
 namespace MackySoft.Ucli.Tests.Cli;
 
@@ -8,14 +8,58 @@ public sealed class CliCommandRegistrationContractTests
 {
     [Fact]
     [Trait("Size", "Medium")]
-    public async Task HelpOutput_ContainsEveryMetadataCommandPath ()
+    public async Task HelpOutput_CommandPathsMatchCommandCatalog ()
     {
         var result = await CliProcessRunner.RunCommand("--help");
 
         Assert.Equal((int)CliExitCode.Success, result.ExitCode);
-        for (var i = 0; i < UcliCommandMetadataCatalog.CommandPaths.Count; i++)
+        Assert.Equal(
+            UcliCommandCatalog.CommandPaths.Order(StringComparer.Ordinal).ToArray(),
+            ExtractHelpCommandPaths(result.StdOut).Order(StringComparer.Ordinal).ToArray());
+    }
+
+    private static string[] ExtractHelpCommandPaths (string helpOutput)
+    {
+        var commandPaths = new List<string>();
+        var lines = helpOutput.Split('\n');
+        var isCommandSection = false;
+
+        for (var i = 0; i < lines.Length; i++)
         {
-            Assert.Contains(UcliCommandMetadataCatalog.CommandPaths[i], result.StdOut, StringComparison.Ordinal);
+            var line = lines[i].TrimEnd('\r');
+            if (!isCommandSection)
+            {
+                isCommandSection = string.Equals(line.Trim(), "Commands:", StringComparison.Ordinal);
+                continue;
+            }
+
+            if (string.IsNullOrWhiteSpace(line))
+            {
+                break;
+            }
+
+            if (!line.StartsWith("  ", StringComparison.Ordinal))
+            {
+                break;
+            }
+
+            commandPaths.Add(ExtractCommandPath(line));
         }
+
+        return commandPaths.ToArray();
+    }
+
+    private static string ExtractCommandPath (string helpLine)
+    {
+        var trimmed = helpLine.TrimStart();
+        for (var i = 0; i < trimmed.Length - 1; i++)
+        {
+            if (trimmed[i] == ' ' && trimmed[i + 1] == ' ')
+            {
+                return trimmed[..i];
+            }
+        }
+
+        throw new InvalidOperationException($"Command help line does not contain a description separator: {helpLine}");
     }
 }
